+++ OnePy/__init__.py
@@ -4,3 +4,6 @@ from portfolio import *
 from strategy import StrategyBase
 from execution import *
 from feed import *
+
+from utils.py3 import with_metaclass
+from utils.metabase import MetaParams